Implement file lock for aperture photometry operations

+++ app/host/transient_tasks.py
@@ -37,6 +37,7 @@ from host.object_store import ObjectStore
 from host.base_tasks import TaskRunner
 from host.base_tasks import get_image_trim_status
 from django.conf import settings
+from pathlib import Path
 
 """This module contains all of the TransientTaskRunners in blast."""
 
@@ -565,10 +566,28 @@ class GlobalApertureConstruction(TransientTaskRunner):
                 object_key = os.path.join(settings.S3_BASE_PATH, local_fits_path.strip('/'))
                 s3.download_object(path=object_key, file_path=local_fits_path)
             assert os.path.isfile(local_fits_path)
+            # Create a lock file to prevent concurrent processes from deleting the data file prematurely
+            lock_path = f'''{local_fits_path}.GlobalApertureConstruction.lock'''
+            Path(lock_path).touch(exist_ok=True)
+            assert os.path.isfile(lock_path)
+            # Construct aperture
             image = fits.open(local_fits_path)
-            aperture = construct_aperture(image, transient.host.sky_coord)
-            # Delete FITS file from local file cache
-            os.remove(local_fits_path)
+            err_to_raise = None
+            try:
+                aperture = construct_aperture(image, transient.host.sky_coord)
+            except Exception as err:
+                err_to_raise = err
+                pass
+            finally:
+                os.remove(lock_path)
+                if not [Path(local_fits_path).parent.glob('*.lock')]:
+                    try:
+                        # Delete FITS file from local file cache
+                        os.remove(local_fits_path)
+                    except FileNotFoundError:
+                        pass
+                if err_to_raise:
+                    raise err_to_raise
             choice += 1
         if aperture is None:
             return self. _failed_status_message()
@@ -645,14 +664,20 @@ class LocalAperturePhotometry(TransientTaskRunner):
         cutouts = Cutout.objects.filter(transient=transient).filter(~Q(fits=""))
 
         for cutout in cutouts:
-            # Download FITS file local file cache
-            s3 = ObjectStore()
             local_fits_path = cutout.fits.name
-            object_key = os.path.join(settings.S3_BASE_PATH, local_fits_path.strip('/'))
-            s3.download_object(path=object_key, file_path=local_fits_path)
+            if not os.path.isfile(local_fits_path):
+                # Download FITS file local file cache
+                s3 = ObjectStore()
+                object_key = os.path.join(settings.S3_BASE_PATH, local_fits_path.strip('/'))
+                s3.download_object(path=object_key, file_path=local_fits_path)
             assert os.path.isfile(local_fits_path)
-            image = fits.open(local_fits_path)
+            # Create a lock file to prevent concurrent processes from deleting the data file prematurely
+            lock_path = f'''{local_fits_path}.LocalAperturePhotometry.lock'''
+            Path(lock_path).touch(exist_ok=True)
+            assert os.path.isfile(lock_path)
 
+            image = fits.open(local_fits_path)
+            err_to_raise = None
             try:
                 photometry = do_aperture_photometry(
                     image, aperture.sky_aperture, cutout.filter
@@ -677,13 +702,19 @@ class LocalAperturePhotometry(TransientTaskRunner):
                     data["magnitude_error"] = photometry["magnitude_error"]
 
                 self._overwrite_or_create_object(AperturePhotometry, query, data)
-                try:
-                    # Delete FITS file from local file cache
-                    os.remove(local_fits_path)
-                except FileNotFoundError:
-                    pass
-            except Exception:
-                raise
+            except Exception as err:
+                err_to_raise = err
+                pass
+            finally:
+                os.remove(lock_path)
+                if not [Path(local_fits_path).parent.glob('*.lock')]:
+                    try:
+                        # Delete FITS file from local file cache
+                        os.remove(local_fits_path)
+                    except FileNotFoundError:
+                        pass
+                if err_to_raise:
+                    raise err_to_raise
         return "processed"
 
 
@@ -733,14 +764,19 @@ class GlobalAperturePhotometry(TransientTaskRunner):
                 break
         query = {"name": f"{cutout_for_aperture.name}_global"}
         for cutout in cutouts:
-            # Download FITS file local file cache
-            s3 = ObjectStore()
             local_fits_path = cutout.fits.name
-            object_key = os.path.join(settings.S3_BASE_PATH, local_fits_path.strip('/'))
-            s3.download_object(path=object_key, file_path=local_fits_path)
+            if not os.path.isfile(local_fits_path):
+                # Download FITS file local file cache
+                s3 = ObjectStore()
+                object_key = os.path.join(settings.S3_BASE_PATH, local_fits_path.strip('/'))
+                s3.download_object(path=object_key, file_path=local_fits_path)
             assert os.path.isfile(local_fits_path)
-            image = fits.open(local_fits_path)
+            # Create a lock file to prevent concurrent processes from deleting the data file prematurely
+            lock_path = f'''{local_fits_path}.GlobalAperturePhotometry.lock'''
+            Path(lock_path).touch(exist_ok=True)
+            assert os.path.isfile(lock_path)
 
+            image = fits.open(local_fits_path)
             # make new aperture
             # adjust semi-major/minor axes for size
             if f"{cutout.name}_global" != aperture.name:
@@ -812,11 +848,13 @@ class GlobalAperturePhotometry(TransientTaskRunner):
                 err_to_raise = err
                 pass
             finally:
-                # Delete FITS file from local file cache
-                try:
-                    os.remove(local_fits_path)
-                except FileNotFoundError:
-                    pass
+                os.remove(lock_path)
+                if not [Path(local_fits_path).parent.glob('*.lock')]:
+                    try:
+                        # Delete FITS file from local file cache
+                        os.remove(local_fits_path)
+                    except FileNotFoundError:
+                        pass
                 if err_to_raise:
                     raise err_to_raise
